Добавление слайсов
