+++ src/store/data-process/data-process.ts
@@ -1,19 +1,19 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { NameSpace } from '../../const';
 import { DataProcess } from '../../types/types';
-import { fetchAddReviewAction, fetchAroundOffersAction, fetchOfferAction, fetchOffersAction, fetchReviewsAction } from '../api-actions';
+import { fetchAddReviewAction, fetchAroundOffersAction, fetchOfferAction, fetchReviewsAction, fetchOffersAction } from '../api-actions';
 
 const initialState: DataProcess = {
   offers: [],
+  isOffersDataLoading: false,
   aroundOffers: [],
   reviews: [],
   offer: null,
   favorites: [],
-  isOffersDataLoading: false,
   error: null
 };
 
-export const userProcess = createSlice({
+export const dataProcess = createSlice({
   name: NameSpace.Data,
   initialState,
   reducers: {
@@ -25,21 +25,8 @@ export const userProcess = createSlice({
   },
   extraReducers(builder) {
     builder
-      .addCase(fetchOffersAction.fulfilled, (state, action) => {
-        state.offers = action.payload;
-        state.isOffersDataLoading = false;
-      })
-      .addCase(fetchOffersAction.pending, (state) => {
-        state.isOffersDataLoading = true;
-      })
-      .addCase(fetchOffersAction.rejected, (state) => {
-        state.isOffersDataLoading = false;
-      })
       .addCase(fetchAroundOffersAction.fulfilled, (state, action) => {
         state.aroundOffers = action.payload;
-      })
-      .addCase(fetchReviewsAction.fulfilled, (state, action) => {
-        state.reviews = action.payload;
       })
       .addCase(fetchReviewsAction.fulfilled, (state, action) => {
         state.reviews = action.payload;
@@ -49,6 +36,18 @@ export const userProcess = createSlice({
       })
       .addCase(fetchAddReviewAction.fulfilled, (state, action) => {
         state.reviews.push(action.payload);
+      })
+      .addCase(fetchOffersAction.fulfilled, (state, action) => {
+        state.offers = action.payload;
+        state.isOffersDataLoading = false;
+      })
+      .addCase(fetchOffersAction.pending, (state) => {
+        state.isOffersDataLoading = true;
+      })
+      .addCase(fetchOffersAction.rejected, (state) => {
+        state.isOffersDataLoading = false;
       });
   }
 });
+
+export const { dropOffer } = dataProcess.actions;
