+++ src/store/user-process/selectors.ts
@@ -1,0 +1,4 @@
+import { State } from '../../types/types';
+import { NameSpace } from '../../const';
+
+export const getAutorisationStatus = (state: State) => state[NameSpace.User].authorizationStatus;
